bugfix: next.js build errors fixed
